Move the dot parsing code in a submodule of dot.
The code is modified to accomodate any graph type that implements Create.
