+++ src/dot.rs
@@ -324,6 +324,59 @@ where
     }
 }
 
+#[cfg(feature = "dot_parser")]
+mod dot_parser {
+    pub use dot_parser::canonical::{Node as CNode};
+    pub use dot_parser::subgraph_free::{AList};
+    use dot_parser::canonical::Graph as CGraph;
+    use dot_parser::ast::Graph as AstGraph;
+    use crate::data::Create;
+    use std::convert::TryFrom;
+
+    pub trait ParseFromDot<'a>: Create<EdgeWeight=AList<(&'a str, &'a str)>, NodeWeight=CNode<(&'a str, &'a str)>> {
+        fn from_dot_graph(dot_graph: CGraph<(&'a str, &'a str)>) -> Self {
+            let node_number = dot_graph.nodes.set.len(); 
+            let edge_number = dot_graph.edges.set.len();
+            let mut graph = Self::with_capacity(node_number, edge_number);
+            let mut node_indices = std::collections::HashMap::new();
+            for node in dot_graph.nodes.set {
+                let ni = graph.add_node(node.1);
+                node_indices.insert(node.0, ni);
+            }
+            for edge in dot_graph.edges.set {
+                let from_ni = node_indices.get(&edge.from).unwrap();
+                let to_ni = node_indices.get(&edge.to).unwrap();
+                graph.add_edge(*from_ni, *to_ni, edge.attr);
+            }
+            graph
+        }
+
+        fn try_from(s: &'a str) -> Result<Self, ()> {
+            let ast = AstGraph::try_from(s).map_err(|_| ())?;
+            let canonical: CGraph<(&'a str, &'a str)> = ast.into();
+            let petgraph = Self::from_dot_graph(canonical);
+            Ok(petgraph)
+        }
+    }
+
+#[macro_export]
+    /// Statically imports a [Graph] from a valid DOT/Graphviz [&str].
+    macro_rules! graph_from_str {
+        ($s:expr) => { ParseFromDot::from_dot_graph(dot_parser_macros::from_dot_string!($s)) };
+    }
+
+#[macro_export]
+    /// Statically imports a [Graph] from a DOT/Graphviz file. The macro expects the file path as argument. 
+    ///
+    /// Notice that, since the graph is imported *statically*, the file must exist at compile time, but
+    /// can be removed at runtime.
+    macro_rules! graph_from_file {
+        ($s:expr) => { ParseFromDot::from_dot_graph(dot_parser_macros::from_dot_file!($s)) };
+    }
+
+
+}
+
 #[cfg(test)]
 mod test {
     use alloc::{format, string::String};
